backend/requirements.dev.txt
aio-pika==9.1.3
alembic==1.9.4
anyio==3.6.2
apscheduler==3.10.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
backoff==2.2.1
boto3==1.26.57
botocore==1.29.57
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.0
fastapi==0.95.2
filelock==3.9.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.15
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
mock==5.0.1
multidict==6.0.4
nodeenv==1.7.0
openpyxl==3.1.1
orjson==3.8.5
packaging==23.0
pamqp==3.2.1
passlib==1.7.4
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.2
pikepdf==8.5.1
Pillow==10.0.1
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.0
psutil==5.9.5
pycparser==2.21
pydantic==1.10.4
pydantic-computed==0.2.2
PyJWT==2.8.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
python3-saml==1.15.0
PyYAML==6.0
rapidfuzz==3.0.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.41.1
sqlalchemy_utils==0.41.1
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
Wand==0.6.11
wrapt==1.14.1
yarl==1.8.2
